Disable subpixel antialiasing for all text
To verify the impact of setting this property, I added "light on dark"
rendering of each typography component to the storybook. I also set the
default text color to off-black (our primary text color) because it's
hard to tell the difference between that and the browser default
(pure black) at a glance.

Test Plan:
npm test

Reviewers: john