Remove default text color from Text component
This was a good idea but in practice it turns out that it prevents
colors from cascading, which is really annoying. This commit reverts
us back to the previous behavior, which was fine. Ideally we'll just
set the default text color in a css reset somewhere instead

Test Plan:
npm test

Reviewers: briangenisio