Update all typography components to be display: block as per the implementation spec (#27)
Update all typography components to be "display: block" as per the implementation spec

Summary:
This diff adds a "display: block" to all components in wonder-blocks-typography
and updates the storybook stories for that component to show that the components
are actually using block style.  It does not modify the Text component which is
still inline.  Lastly, the .babelrc files were updated to transform the object
spread operator.

Reviewers: briangenisio