Whoops, run the pre-commit hooks for #46!
Turns out I was working directly in the webapp submodule, rather than a standalone repo, so pre-commit hooks weren't running.

For this change, I reset and replayed the commit for #46, which ran the appropriate pre-commit hooks, then rebased against `master`. This had the effect of rebuilding the Styleguidst files, in the way my commit should've the first time!