[modal] Create basic StandardModal component
In this change, we implement the basics of the "standard overflow" modal layout, by creating the `StandardModal` component.

I also added a `position: relative` style to our Styleguidist preview areas, to enable the example container to take the full height of the preview area. This means that, when using the resize handle to increase the size of the preview area, the space available to the modal will now correctly increase. (It also introduces an awkward issue, in which decreasing the height of the preview area will add a scrollbar, to help you see the rest of the sizer element that I used to set the preview's initial height. Womp, womp! I think it's a net better experience, since you actually _can_ test height changes now, even though parts of the experience are clunky.)