[modal] Extract out `ModalLauncherPositioner` for testing
Whoops, in #54, I didn't include unit tests for this new closing behavior. In this change, we add tests!

In order to test this functionality more directly, I pulled out `ModalLauncherPositioner` into its own component, so that I could easily render this backdrop and test its click handler without involving other dependencies. If we'd instead tried to test this functionality while baked into `ModalLauncher`, we would've needed to do a full render, including going through the `ModalLauncherPortal`, which would've been messy and liable to fail due to unrelated breakages.

Test Plan:
```
npm run jest
```

And confirm on Styleguidist that launching and closing a modal via backdrop still works as expected.

Additionally, confirm that `ModalLauncherPositioner` does not appear on Styleguidist, because we added a new ignore rule for it.