[modal] Add close button to modal layouts
In this change, we create a `ModalCloseButton` component, and add it to the modal layout components.

This close button doesn't control modal launcher state yet; it just triggers the public callback `onClickCloseButton` on the modal layout component. (We'll probably need to set up some wires via context, in order to control the modal launcher. Coming soon!)

Test Plan:
Open Styleguidist, and take a look at those cute close buttons!
TODO(mdr): Screenshot

Click them, and see an alert, because the Styleguidist examples have an alert handler hooked up to `onClickCloseButton`.

Additionally, use a ruler to confirm that the title and subtitle are indeed centered in the title bar, despite the addition of the close button.

Finally, use the resize handle to confirm that, when the viewport becomes small, the title and subtitle text still wrap naturally.
TODO(mdr): GIF.