[modal] Add close button to modal layouts (#58)
In this change, we create a `ModalCloseButton` component, and add it to the modal layout components.

This close button doesn't control modal launcher state yet; it just triggers the public callback `onClickCloseButton` on the modal layout component. (We'll probably need to set up some wires via context, in order to control the modal launcher. Coming soon!)

Some modals will not want to provide default close behavior at all, but that's not part of the scope of this change. (See WB-134.)

Test Plan:
Open Styleguidist, and take a look at those cute close buttons!
![StandardModal screenshot](https://user-images.githubusercontent.com/59218/39440118-c3667986-4c5e-11e8-8a9a-b23ed83f8036.png)
![TwoColumnModal screenshot](https://user-images.githubusercontent.com/59218/39440121-c9b287d0-4c5e-11e8-8038-f855f2aa05a9.png)

Click them, and see an alert, because the Styleguidist examples have an alert handler hooked up to `onClickCloseButton`.
<img width="916" alt="Alert screenshot" src="https://user-images.githubusercontent.com/59218/39440145-d7c3505c-4c5e-11e8-9adc-a9df04d7b633.png">

Additionally, use a ruler to confirm that the title and subtitle are indeed centered in the title bar, despite the addition of the close button.

Finally, use the resize handle to confirm that, when the viewport becomes small, the `StandardModal` title and subtitle text still wrap naturally.
<img width="380" alt="Small StandardModal screenshot" src="https://user-images.githubusercontent.com/59218/39440170-eb8366e0-4c5e-11e8-9f20-38accf02651b.png">