[modal] Automatically wire close buttons up to ModalLauncher state
In #58, we added close buttons to the modal layout components.

In this change, we wire them up to `ModalLauncher`'s state automatically.

To do this, we use `React.cloneElement` to add an `onClickCloseButton` prop to the provided `modal`. Our two modal layout components (`StandardModal` and `TwoColumnModal`) both support this prop, so this is the only change necessary to integrate the two layers.

An alternative would have been to require the user to perform this wiring themselves, via function-as-children. This would've been super annoying!

Another alternative would've been to use context. This would've been a bit more complex internally, but grant the API a bit more freedom to wrap the `modal` element in other ways. However, our API doesn't _want_ users to wrap the modal in arbitrary ways: we want them to give us a modal layout component instance, and that's it. So, this extra flexibility wouldn't actually be a win right now - but, if we end up needing that flexibility later, maybe this would be a good direction! (Incidentally, I tried to use Flow types to enforce this constraint, but had trouble and decided to punt.)

Along the way, I narrowed our API to accept a `React.Element` rather than a `React.Node`, in order to simplify our usage of `cloneElement`. This also enabled us to remove some complexity in `ModalLauncherPortal`, which used a wrapper element that's no longer necessary.

Test Plan:
```npm test```

In Storybook, use the `ModalLauncher` example to launch a standard modal and a two-column modal. Confirm that both of their close buttons successfully close the launched modal.