[modal] Trap focus in the modal, and wrap tab button presses
In this change, we ensure that tab focus stays within the modal. If the user attempts to tab beyond the modal's boundaries, their tab focus wraps back around.

The general approach was adapted from the WAI-ARIA dialog behavior example, in which they implemented similar behavior for their modal.

Unfortunately, I couldn't get an automated test working :( I left my WIP version of the test in the suite as a `test.skip`, and performed testing manually.

This new login-form example showcases that maybe we want cleverer auto-focus behavior than #70 gives us. (We auto-focus the "log in" button, but we should probably auto-focus the username field.) Will try something cleverer in a later change!

https://khanacademy.atlassian.net/browse/WB-132

# Test Plan:
In Styleguidist, launch the `TwoColumnModal`, which now has a login form in the right-hand column.

Observe that, as in #70, focus is automatically moved inside the modal.

Then, use Tab and Shift-Tab to rotate through the form fields, form buttons, and modal close button. Observe that they are a cycle: pressing Tab when "Log in" is focused takes you to the modal close button, and pressing Shift-Tab when the close button is focused takes you to "Log in", despite them being at start and end of the modal.