Update View and Text to pass along unknown props (#68)
In this change, we allow `View` and `Text` to accept arbitrary props to pass along to the wrapper HTML element. This enables event handlers like `onClick`, and accessibility features like `id` and `aria-hidden`.

Test Plan:
In Styleguidist, look at the new example in the "Core" section.

Click the words "Click me!", and see an alert appear.

Use a screen reader (I used VoiceOver in Safari), and observe that the screen reader does not recognize the text "This text is hidden from screen readers.".