[modal] Mobile-optimized design for standard modal (#64)
In this change, we add mobile-optimized styles to `StandardModal`. We also tweak the example styles, to more accurately showcase mobile design.

We're not tackling other modal layouts yet; that's WB-130. Instead, we're producing one mobile layout, to showcase overall mobile behavior and get high-level design feedback, without slowing down iteration on the other modal layouts with this additional layer of complexity.

We also don't yet prevent the document from scrolling when this modal is open. That's up next, in WB-149.

(When tweaking the examples, I noticed that callers are currently responsible for a number of decisions, like content area padding on desktop and mobile, which we might want abstractions for. We _don't_ want to  this padding into all modals, because many modals don't need it—but maybe it should be a default padding controlled by a flag, or there should be a `ModalContent` component? Long-term food for thought!)

# Test Plan:
Styleguidist examples for `StandardModal` look great on mobile!

Here it is launched on the page.

Also, confirm that all the examples still look as expected on desktop.