[modal] First draft of ARIA dialog behavior (#70)
In this change, we add basic screen reader support for our modal. In particular, we give it the "dialog" ARIA role, mark its title+subtitle nodes as a label and description respectively, and we add very basic focus management.

I only added an ARIA label and description in the `StandardModal`, since `title` and `subtitle` are its props. For the `TwoColumnModal`, I only applied `role="dialog"`, since identifying the appropriate label is trickier. (We just receive a monolith of content as our props.)

Depends on #69, which enables typography components to accept `id`.

https://khanacademy.atlassian.net/browse/WB-138

# Test Plan:
Using VoiceOver in Safari, open the standard modal in the `ModalLauncher` example. Confirm that focus shifts to its "Close modal" button, and that VoiceOver reports that the user is currently in a dialog named "Title".