[modal] Whoops, override `minHeight` in mobile view
I mentioned this in code review for #64, but forgot to change it! In addition to resetting our `maxWidth` and `maxHeight` rules, we also need to reset our `minHeight`. Otherwise, the modal will overflow the screen when it's <200px tall.

Test Plan:
Launch a `StandardModal`, then manually resize the window to be quite narrow and quite short. Confirm that the modal continues to be full-screen, without overflowing.