[modal] Whoops, override `minHeight` in mobile view (#72)
I mentioned this in code review for #64, but forgot to change it! In addition to resetting our `maxWidth` and `maxHeight` rules, we also need to reset our `minHeight`. Otherwise, the modal will overflow the screen when it's <200px tall.

Test Plan:
Launch a `StandardModal`, then manually resize the window to be quite narrow and quite short. Confirm that the modal continues to be full-screen, without overflowing.
![image](https://user-images.githubusercontent.com/59218/39648272-89c99c40-4f96-11e8-8dad-6841272e7fc1.png)