Handle client side navigation in Button
Summary:
Create a getClickableBehavior() function in wonder-blocks-core that will
return either the regular ClickableBehavior or one wrapped in withRouter
that can be used for client side navigation in Button or other components.
This diff updates Button and Link to use getClickableBehavior and updates
the demo app to use the new clientNav prop on Button and Link.