Allow custom styling for Buttons
Allow width, position, margin, and flex styles to be passed in to Buttons. Styles that are not defined in ButtonCore may also be passed in, though we don't explicitly support this, and will be removing this once we strictly type the styles that are passed in.

Test Plan:
View on styleguidist