Add Spacing v1 package. (#67)
This adds the Spacing package and uses it in the Grid package. I'll use it in Modal once the dust settles there, as well. Accessing the sizes by name is a bit clunky - I find it very hard to remember and have to constantly look it up (like I do with the colors) but I'm not of a better solution? Hmm.