Add grid-size support to style property. (#82)
Add in support for passing a function to the style property of our HOC so that we can display certain styles based on the current grid size. This should replace the need for most media queries in our code base (we'll just want to wrap our site in one grid as a base but we can always override it further down).