Re-implement modals to use low-level building blocks. (#81)
Now that we're free from our Grid restrictions it's really easy to build up a series of low-level building blocks to construct our modals! I built a collection of these components to build all of our modals, and to give users the power to construct their own and potentially override specific aspects of the layout, if need be.

Looking at an existing set of modals:

![screenshot 2018-05-09 17 10 56](https://user-images.githubusercontent.com/1615/39840293-e3e35e5e-53ac-11e8-8976-dea94d0c9e99.jpg)

* ModalDialog (red): Holds the entire modal and all of the contents (a set of ModalContentPanes).
* ModalContentPane (orange): Holds a set of ModalTitleBar, ModalHeader, ModalContent, and ModalFooter and renders them all. Additionally is responsible for rendering the ModalCloseButton.
* ModalTitleBar (green): Renders the Title and Subtitle.
* ModalHeader (teal): Renders the section displayed beneath the titlebar and above the contents.
* ModalContent (purple): The main contents of the ModalContentPane (and the only required part of the content pane).
* ModalFooter (blue): An optional footer to display beneath the contents.

This diff also implements the following new additions to the existing modals:
* OneColumnModal is now available!
* StandardModal Headers are now available (they are dark and if provided cause the titlebar to turn dark too).
* StandardModal Preview (displayed next to the main contents and have less padding than a normal contents).

I think you'll find that by breaking everything up all of the components are much simpler now. All of the main modals are quite small and (IMO) easy to understand. If someone wanted to build a similar one they could easily copy the contents of an existing modal and modify it.

Not done yet and will do in a following diff:
* Get mobile support working.
* Write documentation for all of this!