[Link v1] Create LinkCore
In this PR, we create the LinkCore component that will act as a stateless component which renders a state of Link. In a separate PR, we will add ClickableComponent to Link.

Due to security measures around styling the `:visited` pseudo class, we will be adding Visited Link styles in a separate PR.

Test Plan:
View on Stylguidist