[IconButton v1] Create IconButton
Create the `IconButton` component using the same `ClickableBehavior` + `_Core` pattern as `Button` and `Link`.

Note: Since we do not yet have an `Icon` component, we are rendering a plain SVG element, which assumes the size of the SVG to be 9.8px by 9.8px (which is the size of the icon used in the examples).