Add mobile support to modal dialogs. (#84)
This adds in mobile support for all of the modal dialogs. Some of these were easier to implement than other (like with the standard modal). The one and two column modals had to be restructured so that their footer would be positioned outside of the ModalContentPane so that everything else could be scrollable (content, close button, etc.)

I also went in and stripped out a lot of old Grid modal stuff that we're not using anymore (but we are using some of it!) I think there may be room for a new Grid component to make some of this process easier so that we don't have to use React Context but I'll save that for a future diff.

It's probably easiest to just look at the live example on Netlify to see how it works. The non-mobile version should work the same - and the mobile one should match what's proposed in Zeplin.

* All modals should consume the entire viewport, nothing behind it should be visible.
* The horizontal margins are tighter (16px) and the vertical are the same (32px).
* If there is no title bar (and always in one and two-column modals) the "X" is positioned above the rest of the content and scrolls with the content.
* All content, except for the title bar and the footer, scroll off of the screen.
* The preview area in the standard modal disappears.
* The sidebar in the two-column modal becomes a pseudo-header.