Rename Grid to MediaLayout. Add HOC for MediaLayoutWrapper. (#114)
This is a rewrite to change Grid to become a more generic "MediaLayout" component (housed in `wonder-blocks-core` instead of inside grid). There are more improvements that I would like to make but this is a good place to start.

The main changes:

* Moving grid.js out of the Grid package into Core, named media-layout.js.
* Renaming many of the Grid-centric things as "MediaLayout" or "Media" instead (largely automated).
* Adding a HOC to wrap a React Component to give it access to `mediaSize` and `mediaSpec` (making it so that components no longer need to know about React Context to get access to this information).

Looking at the components in styleguidist I can confirm that the Grid and Modal work the same as before.

After having implemented it and converted over to it I'm a bit mixed on using a HOC for this - I think I'd rather have a dedicated component for this (e.g. `<Layout>{({mediaSize}) => ...}</Layout>`).

I was actually thinking that the ideal version of this might be something that behave similar to the existing Grid components and also has the ability to something like `<Layout large={true}>...</Layout>` if you wanted to show something only at large layouts.

Additionally, we have support for layouts in style like so: `style={(mediaSize) => ...}` but I think I'd like to modify it so that we get access to the properties of the spec so you can do something like: `style={({marginSize}) => ({marginTop: marginSize})}` right now you have no way to get this info. Let me know what your thoughts are on these changes and I can make them in a follow-up PR.