update .circleci/config.yml to actually lint our code and run the tests and coverage (#226)
This PR replaces travis-ci with circleci.  Circleci will run on PRs or pushes but prevents duplicate builds when both occur in the case of PRs for branches that aren't in a fork.  It also allows multiple jobs to be run in parallel.  In this case we're running eslint and flow in one job called `lint` and building the code and running the tests along with coverage is in another job called `test`.