WB-338: exclude all external deps from built files (#258)
This is still kind of error prone b/c I had to check things manually.  I would like to create a script to verify that all require/import state for an external dep appear in the modules "dependencies" or "peerDependencies".  That'll be a separate PR b/c I want to get some of these components published.  For now we can verify what's being bundled by looking at the output of webpack which we're now passing `--display-modules` to which will show all modules.  Ones that aren't bundled are marked as `external` and only occupy 42 bytes in the bundle.