Make client-side navigation the default. (#261)
If a React Router is available then we would be defaulting to using client-side navigation and force the dev to opt out of it explicitly (rather than the other way around). This simplifies that logic for using these components a bit. Simply: If a React Router exists (and the href is a non-external URL) then it'll try to use React Router. If you don't want to use React Router then you opt out of it by providing a `directNav={true}`. This replaces the use of `clientNav` (which was the opposite, opting in).

There is similar logic in ActionItem, Link, Button, and IconButton so there is a lot of duplicate code in this diff, but it's all pretty similar.