Enforce style types for all `style` props. (#262)
When working on implementing the new <Layout/> component I realized that I was going to need to have true, enforced, Flow types for the CSS properties that are in Aphrodite and in all of the `style` props.

I've added a new `StyleType` that we should use for all of our `style` props in every component (it's exported from wonder-blocks-core).

Additionally, I've updated the Aphrodite flow rules to match what we're expecting. In the process I confirmed an error that I was seeing when running Jest (that we should've named `::-moz-osx-font-smoothing` to `::MozOsxFontSmoothing`, React didn't like the former). There were a couple small issues like this that I've resolved.

I held off on implementing the exact style props that we want for each component, we can tackle that later, at the very least right now we can enforce that the style props are actually real CSS properties.

Note that I'm not going to land this until `<Layout/>` is ready as I've removed the function option from `<View/>`'s `style` prop. I'll bring that forward in a follow-up PR so that we can evaluate that work separately.