Adding eslint rules and moving shared committed
In this commit, the shared visibility modifier code for use with PopperJS is moved to a new folder called `shared-unpackaged` with an explicit readme to explain what is in there and why code should never go there.

I have also added eslint rules to control 3 specific circumstances:
- Imports from `shared-unpackaged` should never occur
- Imports from `utils` should only occur from test.js files
- Imports within a package should not include files from another package

To support that last one, given that we may add and remove more packages, I have added a script that runs during linting which updates the .eslintrc configuration to ensure all packages are covered.