Rework mock to get past jest shortcoming with ES-style modules
The getType checks that are done when determining the metadata of an automock don't support Module as a type and as such, the automocking fails. If the jest-mock module is hacked to support this type as though it were an object, all is fine, so we should look into updating jest with that, assuming that hack is the right fix.