Use Babel 7 to transform our code. (#306)
* Use Babel 7 to transform our code.

* get jest working with babel 7, kind of

* add @babel/plugin-syntax-dynamic-import

* Get things closer to passing.

* Rework mock to get past jest shortcoming with ES-style modules

The getType checks that are done when determining the metadata of an automock don't support Module as a type and as such, the automocking fails. If the jest-mock module is hacked to support this type as though it were an object, all is fine, so we should look into updating jest with that, assuming that hack is the right fix.