Fix the dropdown menu for mobile.
Pretty easy fix. the bug was that the `touchend` event would happen on
the document, but we weren't checking the menu div (which in rendered
in a portal).
It didn't manifest on desktop, because the "click" event would be
handled by the OptionItem before it got to the document.

Test Plan:
View the wonderblocks documentation with mobile emulation turned on, see
that the <SingleSelect> and <MultiSelect> examples work.