Rename NoSSR to WithSSRPlaceholder [BREAKING CHANGE]
`NoSSR` is a misleading name, because with the `placeholder` param
it's not actually "no" ssr, it's "different" ssr.  (The name is doubly
misleading, since to me "No ssr" means that you can't render this
component during ssr -- it will raise an exception or some such --
when really what it means is that, by default, its children are
ignored when rendered via ssr.)

The new name, I think, makes the intent clearer.  I also made the
placeholder arg required, since I don't think it's right to privilege
"don't show anything" as the default behavior, since it's probably
not the right thing to do most of the time.  (If anything, I'd have
the default be "show a spinner".)

Test Plan:
yarn test