Make mockOnFirstRender required for UniqueIDProvider.
The (old) default behavior of UniqueIDProvider is that it would not
actually render its children during server-side rendering -- that is,
for the initial render.  That seems like a very surprising default to
me, and I think it would be for others as well.

I could change the default, but then you get a different non-obvious
behavior, that you get two different unique-id factories for the
initial render and subsequent renders.

So instead, I just make clients be explicit about which alternative
they prefer.  This helps ensure that they understand the consequences
of their choice.

Test Plan:
yarn test