Button: Prevent user-selection during press on safari mobile
When accessing https://wonder-blocks.netlify.com/ on mobile safari on iOS,
pressing and holding on the default coloured buttons results in the
text-selection UI popping up after a couple seconds. This prevents the
user from being able to click the button.

This commit moves the `user-select: none` styling from the text inside
ButtonCore to the ButtonCore button element itself, which prevents this
misbehaviour on iOS.

Test plan:

1. Ran `yarn start`
2. Opened the resulting page on my iPhone from the local network
3. Pressed and held on the default blue buttons on my phone
    * Verified that no selection popup occurred
    * Verified that releasing the press triggered the modals
4. Ran `yarn jest` and verified that the snapshot changes consisted
   entirely of changes to `user-select: none` styles.