Add onOpen and onClose callbacks to dropdowns
I'm working on a feature that I think requires knowing whether the dropdown is open or closed. So, I tried adding these callbacks. What do you think?

(Specifically, I'm building filters for the district dashboards, and we're tracking which values are selected in the URL's query string. If the user makes multiple changes to the dropdown in a row, i.e. they check a bunch of boxes at once, I want to treat that as an atomic action and use `replaceState`, to avoid creating a new history entry for each individual box they checked. But, once the user closes the dropdown, I want to "finalize" their input and make a permanent history entry for it.)