Add Alexjs to detect insensitive language
This was added to React Native recently and I thought it was a good idea!
https://github.com/facebook/react-native-website/pull/1337

More information: https://alexjs.com/

I think the one piece of our API we could consider changing would be "disabled", but I have deferred that for now.

This feels like a pretty harmless linter and is easy to configure as well, may be worth a shot giving it a try!

Test plan:
I ran `yarn alex` and it passed!