Allow buttons to truncate if parent container is too long
Whilest working on a banner in webapp, I noticed that WB buttons don't correctly truncate when the parent container is too narrow. The issue is the `display: "flex"` styling. Just removing it fixes it, but I chose to instead change it to `display: "inline-block"` so it's clear that it's not set to flex for a reason. I'm happy to remove that and leave the comment if you all feel that would be better.

I also updated the docs so now there's an example of that truncation and we can easily see if it breaks again!

Test plan:
Navigate to `/#button-1` in the WB docs,
verify there's now a section under "Best practices" with an example where the parent is cutting off one of the buttons,
and verify that the button that's cut off is truncating properly with ellipsis.