Fix disjoint union prop types (#624)
* Fix disjoint union prop types

Summary:
Flow gets confused when an optional prop exists in multiple variants of a
disjoint union.  This diff ensures that optional props only exist in a single
variant and that they are not optional in that variant.  To keep flow happy
we need to provide default values for these props when destructuring.

Test Plan:
- load localhost:6060
- see that OnePaneDialog is working correctly

Reviewers: #fe-infra

* commit all code in the buffer