[WB-801] Make sure errors invalidate cache
Here we change how the `invalidateCache` call works and its default
implementation.

1. We now pass the current `CachedEntry` so that invalidation can be decided based on the current value
2. The default implementation now returns that the cache should be invalidated if the cached entry is an errors

This has the effect of causing the client-side to re-request data that
error'd when requested server-side. This is great in that we rehydrate
with the error that the SSR rendered, but a new request to update the
data is made.

We also update the `clone` call on our `ResponseCache` class to be
`cloneCachedData` to make it clearer what we're cloning.