Remove last vestiges of vite
Summary:
vite was an experimental visual regression testing tool that I was working on.
I've been experimenting with chromatic in a different project and I'm pretty
happy with it.  I think we should give it a try in wonder-blocks so that we
can avoid regressions like the one we had for ellipsized labels in buttons a
while ago.  Adopting chromatic will happen in a future PR.

Test Plan: n/a