Set up Chromatic visual regression testing too
Summary:
This diff sets up Chromatic using Github Actions.  Although we're using
CircleCI for most of our CI related needs, I think that we want to use
Github Actions for more things and this is a move in that direction.

Test Plan: See that Chromatic check appears and that the screenshots match the current reference images