Fix issues from #1047
Summary:
I didn't put the 'waiting' logic for the server-side nav case in the right location.  This PR fixes that.

One of the things from the ADR was wanted to make sure people are using 'beforeNav' and 'safeWithNav' only when 'href' is set.  This PR also fixes that by changing the Button props to be a disjoint union.

Issue: WEB-2990

Test plan:
- yarn jest