WEB-2990.?: Use promises instead of async/await
Summary:
Even though we're using async/await in some places in webapp, the code we generator
here doesn't see webapp's version of regeneratorRuntime.  This PR changes the use
of async/await in clickable-behavior.js to Promises instead.

This PR also removes the changes I made to the styleguidist config to drop the import
of @babel/polyfill which was providing regeneratorRuntime locally.

Issue: WEB-2990

Test plan:
- yarn start
- load localhost:6060 and see that it loads fine without regeneratorRuntime errors
- yarn test

Reviewers: #fe-infra