Avoid ambiguous cases in Button's prop types
Summary:
After deploying to webapp, a bunch of flow errors were triggered because the flow types for Button's props are ambiguous when only 'href' is being passed.

The error only triggers when passing the result of a function call instead of just passing a string.  I've added a flow test that reproduces the issue.

This PR fixes it by splitting up the possible cases to be more explicit to avoid ambiguous cases.

Issue: none

Test plan:
- yarn flow

Reviewers: #fe-infra