Add 'xlarge' size variant to Button (#1062)
The reason for adding this variant is that it's currently impossible to create type-safe wrappers around button with the current version of flow that we're using in webapp.  Until we can update flow to a version that fixes this issue we'll have to make do with a "xlarge" size variant in wonder-blocks itself.  The reason why we're skipping over "large" is that mobile has a button size with height: 48px.  We don't use that size anywhere on web but we'd like to align the web version of wonder-blocks with the mobile version as much as possible.  We can always add the "large" size at a later date.

Test Plan:
- look at the buttons on the netlify preview page see that they look decent
- hover over all button sizes with and without icons to see that the width doesn't change

<img width="370" alt="Screen Shot 2020-10-12 at 12 48 07 PM" src="https://user-images.githubusercontent.com/1044413/95771514-2eca6100-0c89-11eb-9a16-5165c9ded54b.png">

Note to reviewer:
- You'll have to review changes in Chromatic before this can proceed, see list of checks below.

Author: kevinbarabash

Reviewers: jeresig, jandrade, russell-ka

Required Reviewers: 

Approved by: jeresig, jandrade

Checks: ✅ Mixed content, ❓ Redirect rules, ❓ Header rules, ❓ Pages changed, ⏭  Auto-approve dependabot PRs, ✅ chromatic

Pull request URL: https://github.com/Khan/wonder-blocks/pull/1062