Implement 'caret' prop for Link
Summary:
This was work we never got to.  I'm doing this now because I want to replace a ClientLink in webapp with a wonder-blocks-link Link but it has a one-off caret and I want to do it the right way.  This will also make updating other links with carets in webapp easier in the future.

As part of this PR I had to stop processing .md files with lint-staged becuase it was writing to test files at the same time jest was trying to write them and this resulted in test failures.

Issue: none

Test plan:
- yarn test
- look at the examples in styleguidist and storybook

Reviewers: #fe-infra