Add 'className' prop type to commonly used components
Summary:
In #1072 I added support for 'className' to 'addStyle', but in order to use 'className' on components without flow complaining we also have to add it to the prop types.  For some of the components, the new prop needed to be piped through manually to a child component.

Issue: none

Test plan:
- yarn start
- in one of the examples for Link, give one of the Links a className, see that it shows up in the inspector
- yarn flow

Reviewers: #fe-infra