WB-1004: Update Button to work with <form>s
Summary:
In the process of adding beforeNav and safeWithNav props, I added a call to e.preventDefault() in response to click, keyboard, and touch events. This unfortunately cause forms with submit buttons to stop working.

Now we only call e.preventDefault() when there's an 'href' since the only default action we need to prevent is navigation.  Also, we only render a 'button' when 'href' isn't set and 'type="submit"' only makes sense when the underlying component is a 'button'.

Issue: WB-1004

Test plan:
- yarn test

Reviewers: #fe-infra