WB-1043 - Fix Dropdown list scrollbar padding
Fixed a wrong width calculation with the `wonder-blocks-dropdown` list
which was causing a padding issue with the scrollbar (when the list was
large enough for the scrollbar to show). Also fixed an issue with the
single-select dropdown highlighting the incorrect selected item when the
dropdown was re-opened.

Issue: https://khanacademy.atlassian.net/browse/WB-1043

Test plan:
1. Open Styleguidist in browser (or React Storybook)
2. Go to: http://localhost:6060/#!/SingleSelect/17 (Storybook: http://localhost:59197/?path=/story/dropdown-multiselect--custom-styles-opened)
3. See that the dropdown scrollbar is perfectly aligned to the right edge (and also the selected item is correctly highlighted when dropdown is re-opened)