WB-1051 - Add States to TextField
Integrated the States that TextField will use. This includes `enabled`,
`disabled`, `error`, and `focused`. Included the `focused` style for
when the user is focused on (editing) the TextField. I attached images
below of this new style.

Issue: https://khanacademy.atlassian.net/browse/WB-1051

Test plan:
1. View the generated Storybook link. Check the "Form / TextField" section.
2. You can also view Styleguidist under the "Form" section.
NOTE: No props have been added yet so the TextField is stuck in the
`enabled` state, but it will also show the `focused` state when you click
on the TextField.