[WB-498] Fix obscurity check for when popper can obscure the anchor (#1130)
## Summary:
Firstly, this adds flow types to popper.js. From that, I fixed an edge case
where the anchor element passed to `isObscured` could be just a reference
object - I don't know under what circumstances that occurs, though I had seen
an error along those lines in the past, so that should be fixed if it happens
anywhere.

Finally, and most importantly, I have fixed the issue that would cause the
popper (whether a tooltip, menu, dropdown, or popover) to flicker or not appear
when being scrolled near the top of the screen. It turns out this was because
the popper portion was coming back as something covering the anchor element.
To work around this, the popper (or child of a popper, as was the case with
the action menu) is set to not receive pointer events while the obscurity check
is completed.

Issue: WB-498

## Test plan:
With the test site, you can compare with "prod".
In this video, you can see the behavior in production that clearly illustrates
the problem being fixed.

https://user-images.githubusercontent.com/1266297/120145758-a9f56e80-c1a9-11eb-9879-131406e755a6.mov

Using the test site, one can see that this no longer occurs.

https://user-images.githubusercontent.com/1266297/120146006-09ec1500-c1aa-11eb-853f-922662992445.mov



Author: somewhatabstract

Reviewers: somewhatabstract, jeresig

Required Reviewers: 

Approved by: jeresig

Checks: ✅ Mixed content, ⚪️ Redirect rules, ⚪️ Header rules, ⚪️ Pages changed, ⏭  Auto-approve dependabot PRs, ✅ chromatic

Pull request URL: https://github.com/Khan/wonder-blocks/pull/1130