WB-808 - Add right padding to ModalHeader title
Added right padding to the `title` of `ModalHeader` due to it
overlapping with the close button of the `OnePaneDialog`.
The Jira ticket specifically mentioned that the "Change Email Modal"
would show this issue on a mobile phone. The "Change Email Modal" it was
referring to was involving the `ModalHeader`. The screenshots below show
the before and after with the best attempt at recreating the issue in
Styleguidist.

Issue: https://khanacademy.atlassian.net/browse/WB-808

Test plan:
1. Open Styleguidist.
2. Go to the "Modal" section and choose any modal with a close button.
3. Edit the example and add a long title.
4. You can play around with the screen size and see how the title now
   wraps before hitting the boundaries of the close button.