WB-1065 - Add States to LabeledTextField
Added the `error` and `focused` state to the `LabeledTextField`. The
`enabled` state was added in the form of a `disabled` prop where it is set to
`false` by default. Included unit tests to cover these new states and prop
(except the `error` state because it will be covered once validation is added).

Issue: https://khanacademy.atlassian.net/browse/WB-1065

Test plan:
1. View the code to see the new states and prop
2. View Styleguidist and / or React Storybook examples under Form / LabeledTextField
   (see generated links)