Update dropdown-core to focus the menu container (element with role "menu" or "lostbox" that contains the menu items) when the menu is opened
Summary:
This is so that mobile screen readers can access the menu. Without this change, focus is not shifted into the menu when it is opened with a screen reader, and screen reader users cannot interact with the menu.

Issue: LP-10008

Test plan:
yarn jest

Open a dropdown menu with a mobile screen reader (VoiceOver/Safari/iOS or TalkBack/Chrome/Android)
Note that your focus jumps to the menu container